pull classic updates section into its own file
